Correctly invalidate intermediate vuids
Because of the "&& usp->server_info" test in get_valid_user_struct,
invalidate_vuid() called for an intermediate vuid would never do what it was
supposed to do. There is no server_info in the intermediate vuids.

This fixes a memleak, it was found for a client that does silly sequences of
sesssionsetup/ulogoff for every operation.