Add a function libnet_smbconf_get_config() to libnet_conf.c
This gets the whole config as a set of lists (of share names
and corresponding lists of parameter names and values). The function
is an aggregate of libnet_smbconf_get_share_names() and
libnet_smbconf_getshare().

Michael
(This used to be commit 94e97a72548a7f76a5273346d472e3ba5b24795a)