Ensure that winbindd_getgroups() can deal with a UPN name.
A user logging in via GDM was not getting a complete list of supplementary
groups in his/her token.  This is because getgroup() was not able to
find the winbindd_domain* using the DNS name.  Fallback to matching the DNS
name is the short name match failes.
(This used to be commit 2030a8de19a2c7c735a8aa367dd953e4a5c447b8)