From Alexander Bokovoy
names is a 'const char **names' here, passed from upper level. Passing
'names' as a talloc context will break things as 'names' is usually
likely is a local variable, not a talloc'd. And we want to have array
elements on a context of the array anyway.

Jeremy.
(This used to be commit 22a4129fabf70e01cc638c2f7347c28201838830)