Fix stat results to be consistent between smbc_stat and smbc_fstat.
We create a kludged inode based on the checksum of the path.  We therefore
need to use the same (full) path when calculating it in both smbc_stat() and
smbc_fstat().

If struct stat has an rdev field, set it to zero.

Derrell
(This used to be commit b4282fbd6d27d868b2d5c04bb72d2d7421822da1)