Fix a segfault
Found by the IBM checker.

Jeremy, you changed this from 'dirname == 0' which does not make sense at all
in the old code to '!dirname == 0' which from my point of view makes even less
sense, because the precedence of ! is according to Harbison/Steele higher than
the precedence of ==. Please check that the code now actually does what it's
supposed to do.

Thanks!
(This used to be commit c8def5564739c4c754001f84fcfb77f1093b96c3)