Fix open file tracking in vfs_cifs so that oplock breaks can propagate
Oplock breaks were not propagating because the list of open files was not
being maintained.

This fixes that based on best-guess of how it should work.

It has been tested manually with windows XP client obtaining an oplock from a
windows 2003 server, which then broke the lock when smbclient read the same file.

Previously the smbclient read blocked until the oplock timed out