pidl/Samba4::NDR::Parser: correctly get the name of an array element
When we have "*r->out.ous"
(char ***ous, a pointer to a pointer to an array of pointers).
we need to use "(*r->out.ous)[3]" to access the 3rd
element of the array "*r->out.ous[3]" was generated before,
but that's the same as "*(r->out.ous[3])" which would mean
the array would apply to a different level.

This patch prepares support for:

[out,ref,size_is(,num)] [string,charset(UTF16)] uint16 ***names;

It means a [ref] pointer to a [unique] pointer to an array
of [unique] pointers which point to an UTF16 string.

metze
(This used to be commit ec0ee2aa5f4bef32f09a426d91c28c985f843038)