NET-API-BECOME-DC: at least find the 'provision' function in python mode
This is still broken, as we call provision() with wrong arguments.
Also the python provision() function, doesn't allow us
to specify the locations of the partition ldb files,
this is strictly needed otherwise the torture test destroys
a production databases.

I think we should have a provision_become_dc() again!

metze