registry: Implement recursive deletes for regf-backed registry.
When deleting a registry key that contains subkeys or values, Windows performs a
recursive deletion that removes any subkeys or values.  This update makes
deletes for an regf-backed registry consistent with Windows.

The regf-backed registry does not have transactional integrity when performing
multiple operations.  Therefore, if an error occurs during the recursive
deletion, the regf-backed registry may be left in an inconsistent state.
(This used to be commit b0321bad290d1a9399b4aba140a622e3af6d7575)