Use consistent hierarchical naming for tests used in selftest.
This is mainly cosmetical and so it would in the future be possible to print a tree of the tests.
People using "make test TESTS=..." may have to specify a slightly different regular expression
than they used to.
(This used to be commit fe0f2173c006573513f54c45c7bbc9d6855494af)