unixinfo.idl: don't use utf8string anymore
Note:

This changes the on wire format!!!

But it's a very bad idea to keep utf8string
as it's not valid NDR encoding and noone
would be able to write a compatible IDL
file using another compiler than pidl.

Maybe we should change the uuid,
but as this isn't really used yet
I keep the old one.

metze
(This used to be commit 5753d7fd8300d36558d3304418e1b0a300f56575)