pvfs_oplock: auto release oplocks after a timeout
Remember that we sent an oplock break to
a client and don't resend. If the filesystem
layer tries to send a new break and the client
has not released after a former oplock break
after the timeout interval, we need to auto release
the oplock.

metze
(This used to be commit bfb0888578677856b2b6b72471f542d0d5d7b838)